[Feat]2차시도 - panzoom
